shiro

+++ src/main/java/com/dongfang/monitor/dao/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends BaseRepository<User, Integer>,JpaSpecificationExecutor<User> {
+public interface UserRepository extends BaseRepository<User, Long>,JpaSpecificationExecutor<User> {
 
     User findByUsername(String username);
